pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "auto-job-apply-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crawlee[all]>=1.4.0",
    "fastapi>=0.133.1",
    "playwright-stealth>=2.0.2",
    "psycopg2-binary>=2.9.11",
    "uvicorn>=0.41.0",
]

[tool.setuptools.packages.find]
where = ["src"]
